# openapi.yaml
openapi: "3.1.0"
info:
  title: Hushclaw API
  version: "1.0.0"
  description: |
    Hushclaw Skill Store public & authenticated REST API.

    ## Authentication

    Protected endpoints accept credentials in either of two ways:

    | Method | Value |
    |--------|-------|
    | HTTP Header | `Authorization: Bearer <token>` |
    | Cookie | `hc_session=<token>` |

    Obtain a token via the **Auth** flow:
    1. `POST /api/v1/auth/send-code` – request a 6-digit OTP by email.
    2. `POST /api/v1/auth/verify-code` – exchange the OTP for a Bearer token (valid 30 days).
servers:
  - url: /api/v1
    description: Current server (relative)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Token obtained from `POST /api/v1/auth/verify-code`
    CookieAuth:
      type: apiKey
      in: cookie
      name: hc_session
      description: Session cookie set after web login

  # ── Reusable schemas ────────────────────────────────────────────────────────
  schemas:
...
    SendCodeRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
    VerifyCodeRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: "482910"
        client_name:
          type: string
          nullable: true
          description: Optional label for the token (e.g. app name or device)
          example: "My CLI Tool"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "hc_abc123..."
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Seconds until expiry
          example: 2592000
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Skill not found"

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /auth/send-code:
    post:
      tags: [Auth]
      summary: Send OTP email
      description: |
        Sends a 6-digit verification code to the provided email address.

        **Rate limit:** One code per 60 seconds. Returns `429` if called again too soon.
        The code expires in 10 minutes (configurable via `OTP_EXPIRE_MINUTES`).
      operationId: sendCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendCodeRequest"
      responses:
        "200":
          description: Code sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Verification code sent. Check your email."
        "422":
          description: Invalid email address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit – please wait 60 seconds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-code:
    post:
      tags: [Auth]
      summary: Verify OTP and obtain token
      description: |
        Exchange a valid 6-digit OTP for a long-lived Bearer token (default: 30 days).

        Use the returned `access_token` as `Authorization: Bearer <token>` in subsequent requests.

        **Failure rules:**
        - After **3 wrong attempts** the OTP is invalidated; request a new one.
        - Expired OTPs are rejected immediately.
      operationId: verifyCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeRequest"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error (e.g. invalid email format)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Account ─────────────────────────────────────────────────────────────────
...